adding courses system
basicially its working but i will add steps structure

more improvements

+++ app/src/main/java/com/refupanker/teachmetech/view/FragmentExplore.kt
@@ -123,7 +123,7 @@ class FragmentExplore : Fragment() {
                     "Hello descr \n $token",
                     "General",
                     Date(),
-                    Random.nextInt(0, 1000)
+                    Random.nextLong(0, 1000)
                 )
             )
         }
